updated proxy
Can now fetch data down to sensor level

+++ InnoviApiProxy/Coordinate.cs
@@ -6,8 +6,8 @@ namespace InnoviApiProxy
 {
     public class Coordinate
     {
-        public int Latitude { get; set; }
-        public int Longitude { get; set; }
-        public int Altitude { get; set; }
+        public float Latitude { get; set; }
+        public float Longitude { get; set; }
+        public float Altitude { get; set; }
     }
 }
